# contracts/openapi/api-v1.json
{
  "components": {
    "schemas": {
...
      "AuditBaselineComparisonV1": {
        "additionalProperties": false,
        "description": "Comparison vs a certified baseline (optional).",
        "properties": {
          "baseline_id": {
            "title": "Baseline Id",
            "type": "string"
          },
          "diff": {
            "additionalProperties": true,
            "title": "Diff",
            "type": "object"
          },
          "matches_baseline": {
            "title": "Matches Baseline",
            "type": "boolean"
          }
        },
        "required": [
          "baseline_id",
          "matches_baseline"
        ],
        "title": "AuditBaselineComparisonV1",
        "type": "object"
      },
      "AuditFormulaV1": {
        "additionalProperties": false,
        "description": "One unique formula recorded by a calculator during the simulation.",
        "properties": {
          "calculator": {
            "title": "Calculator",
            "type": "string"
          },
          "formula": {
            "title": "Formula",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "used_count": {
            "default": 1,
            "minimum": 1.0,
            "title": "Used Count",
            "type": "integer"
          }
        },
        "required": [
          "calculator",
          "formula",
          "stage"
        ],
        "title": "AuditFormulaV1",
        "type": "object"
      },
      "AuditLineageSummaryV1": {
        "additionalProperties": false,
        "description": "Compact summary of the lineage graph.",
        "properties": {
          "nodes_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Nodes Count",
            "type": "integer"
          },
          "roots": {
            "items": {
              "type": "string"
            },
            "title": "Roots",
            "type": "array"
          },
          "stages_summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Stages Summary",
            "type": "object"
          }
        },
        "title": "AuditLineageSummaryV1",
        "type": "object"
      },
      "AuditParametersUsedV1": {
        "additionalProperties": false,
        "description": "Split of inputs consumed by the simulation: request vs parametrization.",
        "properties": {
          "excel_refs": {
            "items": {
              "$ref": "#/components/schemas/LineageRefV1"
            },
            "title": "Excel Refs",
            "type": "array"
          },
          "parametrization": {
            "additionalProperties": true,
            "title": "Parametrization",
            "type": "object"
          },
          "request": {
            "additionalProperties": true,
            "title": "Request",
            "type": "object"
          }
        },
        "title": "AuditParametersUsedV1",
        "type": "object"
      },
      "AuditResponseV1": {
        "additionalProperties": false,
        "description": "Audit envelope for one simulation.\n\nF9 placeholders:\n    engine_version:        literal until WAVE 14 wires the real SemVer.\n    formula_set:           literal until WAVE 14 wires the manifest hash.\n    parametrization_hashes: empty dict today; WAVE 14 will populate.",
        "properties": {
          "api_version": {
            "const": "api-v1",
            "default": "api-v1",
            "title": "Api Version",
            "type": "string"
          },
          "baseline_comparison": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuditBaselineComparisonV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "engine_version": {
            "default": "engine-v2",
            "title": "Engine Version",
            "type": "string"
          },
          "formula_set": {
            "default": "formula-set-v2-7",
            "title": "Formula Set",
            "type": "string"
          },
          "formulas": {
            "items": {
              "$ref": "#/components/schemas/AuditFormulaV1"
            },
            "title": "Formulas",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "lineage": {
            "$ref": "#/components/schemas/AuditLineageSummaryV1"
          },
          "parameters_used": {
            "$ref": "#/components/schemas/AuditParametersUsedV1"
          },
          "parametrization_hashes": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Parametrization Hashes",
            "type": "object"
          },
          "simulation_id": {
            "title": "Simulation Id",
            "type": "string"
          }
        },
        "required": [
          "simulation_id",
          "lineage"
        ],
        "title": "AuditResponseV1",
        "type": "object"
      },
      "AuditSimulationSummaryV1": {
        "additionalProperties": false,
        "description": "Compact summary for the listing endpoint.",
        "properties": {
          "generated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "nodes_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Nodes Count",
            "type": "integer"
          },
          "roots_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Roots Count",
            "type": "integer"
          },
          "simulation_id": {
            "title": "Simulation Id",
            "type": "string"
          },
          "stages": {
            "items": {
              "type": "string"
            },
            "title": "Stages",
            "type": "array"
          }
        },
        "required": [
          "simulation_id"
        ],
        "title": "AuditSimulationSummaryV1",
        "type": "object"
      },
      "AuditValueExplanationV1": {
        "additionalProperties": false,
        "description": "Cadena de lineage humanamente legible para un valor concreto.",
        "properties": {
          "calculator": {
            "default": "",
            "title": "Calculator",
            "type": "string"
          },
          "explanation": {
            "default": "",
            "title": "Explanation",
            "type": "string"
          },
          "formula": {
            "default": "",
            "title": "Formula",
            "type": "string"
          },
          "refs_chain": {
            "items": {
              "$ref": "#/components/schemas/LineageRefV1"
            },
            "title": "Refs Chain",
            "type": "array"
          },
          "simulation_id": {
            "title": "Simulation Id",
            "type": "string"
          },
          "stage": {
            "default": "",
            "title": "Stage",
            "type": "string"
          },
          "value": {
            "title": "Value"
          },
          "value_name": {
            "title": "Value Name",
            "type": "string"
          }
        },
        "required": [
          "simulation_id",
          "value_name"
        ],
        "title": "AuditValueExplanationV1",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "VisionImprimibleApiResponseV1": {
        "description": "Success envelope: ``data`` contains only ``vision_imprimible``.",
        "examples": [
          {
            "data": {
              "vision_imprimible": {
                "analisis_grafico": {
                  "evolucion_ingreso_neto": [
                    {
                      "ingreso_neto": 100000000,
                      "mes": 1
                    }
                  ],
                  "waterfall": [
                    {
                      "concepto": "ingreso_neto",
                      "valor": 100000000
                    }
                  ]
                },
                "comparativo_escenarios": {
                  "escenarios": [],
                  "total": {
                    "componente_fijo_porcentaje": 0,
                    "componente_variable_porcentaje": 1,
                    "modelo_cobro": "Variable",
                    "tarifa_fija": 0,
                    "tarifa_variable": 10000
                  }
                },
                "contingencias_ajustes": [
                  {
                    "aplicado": 0.21,
                    "concepto": "margen_objetivo",
                    "estado": "dentro_rango",
                    "minimo": 0.17
                  },
                  {
                    "aplicado": 0.05,
                    "concepto": "contingencia_operativa",
                    "estado": "dentro_rango",
                    "maximo": 0.08,
                    "minimo": 0.05
                  },
                  {
                    "aplicado": 0.01,
                    "concepto": "contingencia_comercial",
                    "estado": "dentro_rango",
                    "maximo": 0.03,
                    "minimo": 0.01
                  },
                  {
                    "aplicado": 0.02,
                    "concepto": "markup",
                    "estado": "dentro_rango",
                    "maximo": 0.08,
                    "minimo": 0.02
                  },
                  {
                    "aplicado": 0,
                    "concepto": "descuento",
                    "estado": "dentro_rango",
                    "maximo": 0.08,
                    "minimo": 0
                  }
                ],
                "control_aprobacion": {
                  "aprobaciones": [],
                  "criterios_riesgo": [
                    {
                      "categoria": "Cliente",
                      "criterio": "Criterio 1",
                      "id": 1,
                      "nivel": 1
                    },
                    {
                      "categoria": "Cliente",
                      "criterio": "Criterio 2",
                      "id": 2,
                      "nivel": 1
                    },
                    {
                      "categoria": "Cliente",
                      "criterio": "Criterio 3",
                      "id": 3,
                      "nivel": 1
                    },
                    {
                      "categoria": "Cliente",
                      "criterio": "Criterio 4",
                      "id": 4,
                      "nivel": 1
                    },
                    {
                      "categoria": "Cliente",
                      "criterio": "Criterio 5",
                      "id": 5,
                      "nivel": 1
                    },
                    {
                      "categoria": "Cliente",
                      "criterio": "Criterio 6",
                      "id": 6,
                      "nivel": 1
                    },
                    {
                      "categoria": "Cliente",
                      "criterio": "Criterio 7",
                      "id": 7,
                      "nivel": 1
                    },
                    {
                      "categoria": "Cliente",
                      "criterio": "Criterio 8",
                      "id": 8,
                      "nivel": 1
                    },
                    {
                      "categoria": "Cliente",
                      "criterio": "Criterio 9",
                      "id": 9,
                      "nivel": 1
                    },
                    {
                      "categoria": "Cliente",
                      "criterio": "Criterio 10",
                      "id": 10,
                      "nivel": 1
                    }
                  ],
                  "facturacion_mensual_proyectada": 100000000,
                  "nivel_riesgo": {
                    "clasificacion": "Medio",
                    "score_cliente": 1.6,
                    "score_deal": 1.9,
                    "score_operativo": 2.1
                  },
                  "requiere_aprobacion": true
                },
                "economics": {
                  "costo_mensual": 79000000,
                  "escenario_seleccionado": "Escenario 1",
                  "estado_margen": "cumple_minimo",
                  "ingreso_mensual": 100000000,
                  "margen_deal": 0.21,
                  "moneda": "COP",
                  "valor_total_contrato": 1200000000
                },
                "ficha_deal": {
                  "ajuste_precio": {
                    "componente_humano": "IPC",
                    "componente_tecnologico": "20% SMMLV 80% IPC",
                    "frecuencia": "Anual"
                  },
                  "antiguedad_cliente": "Cliente Antiguo",
                  "ciudad": "Bogota",
                  "cliente": "Cliente A",
                  "descuento_aplicado": 0,
                  "fecha_inicio": "2026-07-01",
                  "meses_contrato": 12,
                  "periodo_pago_dias": 30,
                  "porcentaje_imprevisto": 0,
                  "sede": "Toberin",
                  "servicio": "SAC",
                  "tipo_cliente": "Grupo Aval",
                  "volumen_base_mensual": 100000
                }
              }
            },
            "success": true
          }
        ],
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VisionImprimibleDataV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "VisionImprimibleApiResponseV1",
        "type": "object"
      },
      "VisionImprimibleDataV1": {
        "additionalProperties": false,
        "properties": {
          "vision_imprimible": {
            "$ref": "#/components/schemas/VisionImprimibleResponseV1"
          }
        },
        "required": [
          "vision_imprimible"
        ],
        "title": "VisionImprimibleDataV1",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Motor de pricing + API de parametrización y simulación",
    "title": "NEXA Simulator API",
    "version": "1.0.0",
    "x-contract-version": "api-v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/audit/simulation/{simulation_id}": {
      "get": {
        "description": "Return the audit envelope for `simulation_id`.\n\nReturns 404 if the simulation was not run with `with_lineage=True`.",
        "operationId": "get_simulation_audit_api_v1_audit_simulation__simulation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "simulation_id",
            "required": true,
            "schema": {
              "title": "Simulation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Full audit envelope for a simulation.",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/v1/audit/simulation/{simulation_id}/baseline-diff": {
      "get": {
        "description": "Compare KPIs from the persisted lineage vs the baseline's kpis.json.",
        "operationId": "diff_vs_baseline_api_v1_audit_simulation__simulation_id__baseline_diff_get",
        "parameters": [
          {
            "in": "path",
            "name": "simulation_id",
            "required": true,
            "schema": {
              "title": "Simulation Id",
              "type": "string"
            }
          },
          {
            "description": "Certified baseline id.",
            "in": "query",
            "name": "baseline_id",
            "required": true,
            "schema": {
              "description": "Certified baseline id.",
              "title": "Baseline Id",
              "type": "string"
            }
          },
          {
            "description": "Baseline collection version under storage/baselines/.",
            "in": "query",
            "name": "baseline_version",
            "required": false,
            "schema": {
              "default": "v2-7-certified",
              "description": "Baseline collection version under storage/baselines/.",
              "title": "Baseline Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditBaselineComparisonV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare a simulation's kpis vs a certified baseline.",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/v1/audit/simulation/{simulation_id}/explain": {
      "get": {
        "description": "Return the trace_back chain + human explanation for `value_name`.",
        "operationId": "explain_value_api_v1_audit_simulation__simulation_id__explain_get",
        "parameters": [
          {
            "in": "path",
            "name": "simulation_id",
            "required": true,
            "schema": {
              "title": "Simulation Id",
              "type": "string"
            }
          },
          {
            "description": "Canonical value name in the lineage graph.",
            "in": "query",
            "name": "value_name",
            "required": true,
            "schema": {
              "description": "Canonical value name in the lineage graph.",
              "title": "Value Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditValueExplanationV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lineage chain for a specific value in a simulation.",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/v1/audit/simulations": {
      "get": {
        "description": "Return summaries of simulations that have a stored lineage graph.",
        "operationId": "list_audit_simulations_api_v1_audit_simulations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditSimulationSummaryV1"
                  },
                  "title": "Response List Audit Simulations Api V1 Audit Simulations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List simulations with persisted lineage.",
        "tags": [
          "audit"
        ]
      }
    },
...
    "/api/v1/simulation/{simulation_id}/results": {
      "get": {
        "description": "Respuesta pública de una ejecución, limitada a la hoja Visión Imprimible.\n\nEl documento técnico completo permanece persistido para auditoría y para\nlas rutas especializadas, pero no se expone en este endpoint.",
        "operationId": "get_results_api_v1_simulation__simulation_id__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "simulation_id",
            "required": true,
            "schema": {
              "title": "Simulation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisionImprimibleApiResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Results",
        "tags": [
          "simulation-results"
        ]
      }
    },
...
    "/api/v1/simulation/{simulation_id}/traceability": {
      "get": {
        "description": "Trazabilidad contractual persistida para una simulación.",
        "operationId": "get_traceability_api_v1_simulation__simulation_id__traceability_get",
        "parameters": [
          {
            "in": "path",
            "name": "simulation_id",
            "required": true,
            "schema": {
              "title": "Simulation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Traceability",
        "tags": [
          "simulation-results"
        ]
      }
    },
...
  }
}